+++ app/living-scholar/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import LivingScholar from "@/components/living-scholar/LivingScholar";
+
+export default function LivingScholarPage() {
+  return <LivingScholar />;
+}