+++ components/living-scholar/index.ts
@@ -1,0 +1,7 @@
+export { default as LivingScholar } from "./LivingScholar";
+export { default as MorningBrief } from "./MorningBrief";
+export { default as AcademicDNAVisualizer } from "./AcademicDNAVisualizer";
+export { default as OpportunityGalaxy } from "./OpportunityGalaxy";
+export { default as ScholarTimeline } from "./ScholarTimeline";
+export { default as OracleCopilot } from "./OracleCopilot";
+export { default as GrowthScore } from "./GrowthScore";